upgrade: Expose precheck failures to client
When there was some exception during prechecks execution, the info was
stored in the status file but not returned to the client as a response.
Instead a generic 500 response was generated by the framework as a result
of exception propagation.
Now the exceptions are caught in the controller and returned in proper
JSON response.

(cherry picked from commit 379e859039f32839eef372057eff92fbc87cf8a1)