upgrade: dump Monasca, Grafana databases (SOC-9772)
Up until SUSE OpenStack Cloud 8, Monasca ran its own MariaDB
instance on the monasca-server node which hosted Monasca's
metrics database and Grafana's configuration database. As of
Cloud 9 we use the OpenStack cloud's main MariaDB instance for
both of these, hence we need to create a dump of the the data
from both in order to later migrate them to the main MariaDB
instance.