Address review comments and restructure
"restructure": the crowbar-dump-monasca-db.sh script is now simplified to only
dump one database and is deployed to both the monasca-server node and the first
database-server node. The databases being dumped are the Monasca metrics
database and the Grafana database on the respective nodes. This ensures the
dump is later available on the nodes where the recipes restoring the Monasca
metrics database and the Grafana database are being applied.